Rename `src/interfaces.ts` to `src/types.ts`
